Add support for variable reassignment
The reassignment also shows the original variable declaration in the AST, so there will have to be careful consideration around these variables. I plan to add a first-phase optimization pass that 'sweeps up' the duplicate variables and, more generally, all variables that have trivial expressions and that are not a part of expressions themselves. By definition, these would be unused.